apps/api/serializers.py
from rest_framework import serializers
from .models import Post

class PostSerializer(serializers.ModelSerializer):
    class Meta:
        model = Post
        fields = ['id', 'title', 'author', 'is_public', 'content', 'category', 'slug', 'featured_image', 'date_created']

        